hexagon: make SMP work under the VM
Each virtual processor has its own vpid, which vmintop_post needs to
target an IPI, and the base IPI IRQ has to stay clear of the interrupts
devices claim, so make it configurable.  The VM caches translations per
address space and vmclrmap only purges the calling CPU's current one,
so flushes for an mm running elsewhere go out by IPI, and flushes for
an mm that is not running anywhere are deferred to its next switch_mm()
via context.need_invalidate.  A pgd page can be recycled from another
mm, whose ASID state the VM would otherwise resume, so mark a freshly
allocated one for invalidation too.

Suggested-by: Richard Kuo <rkuo@quicinc.com>

+++ arch/hexagon/mm/init.c
@@ -37,14 +37,14 @@ unsigned long highstart_pfn, highend_pfn;
 unsigned long _dflt_cache_att = CACHEDEF;
 
 /*
- * The current "generation" of kernel map, which should not roll
- * over until Hell freezes over.  Actual bound in years needs to be
- * calculated to confirm.
+ * The current "generation" of kernel map.  A single word so that
+ * reads are atomic on a 32-bit machine; switch_mm() compares
+ * generations for inequality, so wrapping is harmless.
  */
 DEFINE_SPINLOCK(kmap_gen_lock);
 
 /*  checkpatch says don't init this to 0.  */
-unsigned long long kmap_generation;
+unsigned long kmap_generation;
 
 void sync_icache_dcache(pte_t pte)
 {
